+++ pi-beacon/src/main/java/com/ibm/pi/beacon/onBootComplete.java
@@ -13,20 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
+package com.ibm.pi.beacon;
 
-package com.ibm.pisdk;
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
 
 /**
- * This interface provides a callback method for the PIAPIAdapter's asynchronous calls.
+ * This receiver will instantiate a beacon sensor on boot up of the device.
+ * If the sensor was running before the user restarted the phone, it will pick up where it left off
  *
  * @author Ciaran Hannigan (cehannig@us.ibm.com)
  */
-public interface PIAPICompletionHandler {
+public class onBootComplete extends BroadcastReceiver {
 
-    /**
-     * Provides the results of the API call.
-     *
-     * @param result result of asynchronous call from API.
-     */
-    void onComplete(PIAPIResult result);
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        PIBeaconSensor sensor = PIBeaconSensor.getInstance(context, null);
+    }
 }
