+++ pi-core/src/main/java/com/ibm/pi/core/PIAPIResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package com.ibm.pisdk;
+package com.ibm.pi.core;
 
 import android.graphics.Bitmap;
