+++ pi-core/src/main/java/com/ibm/pi/core/PILogger.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  **/
 
-package com.ibm.pisdk;
+package com.ibm.pi.core;
 
 import android.util.Log;
