+++ pi-core/src/main/java/com/ibm/pi/core/doctypes/PIBeacon.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  **/
 
-package com.ibm.pisdk.doctypes;
+package com.ibm.pi.core.doctypes;
 
 import com.ibm.json.java.JSONArray;
 import com.ibm.json.java.JSONObject;
