+++ pi-core/src/main/java/com/ibm/pi/core/doctypes/PISensor.java
@@ -1,4 +1,4 @@
-package com.ibm.pisdk.doctypes;
+package com.ibm.pi.core.doctypes;
 
 import com.ibm.json.java.JSONArray;
 import com.ibm.json.java.JSONObject;
