+++ pi-core/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">PI Core</string>
+    <string name="pi_shared_pref">com.ibm.pisdk.shared_prefs</string>
 </resources>
