--- pi-sdk/build.gradle
@@ -1,58 +1,0 @@
-apply plugin: 'com.android.library'
-
-apply plugin: 'com.jfrog.bintray'
-apply plugin: 'com.github.dcendents.android-maven'
-
-// This is the library version used when deploying the artifact
-version = "1.3.0"
-
-ext {
-    bintrayRepo = 'maven'
-    bintrayName = 'pi-sdk'
-
-    publishedGroupId = 'com.ibm.pi'
-    libraryName = 'pi-sdk'
-    artifact = 'pi-sdk'
-    libraryVersion = version
-    libraryDescription = 'A library that allows for easy integration with IBM Presence Insights service available on Bluemix.'
-
-    siteUrl = 'https://github.com/presence-insights/pi-clientsdk-android'
-    gitUrl = 'https://github.com/presence-insights/pi-clientsdk-android.git'
-
-    developerId = 'cehannig'
-    developerName = 'Ciaran Hannigan'
-    developerEmail = 'cehannig@us.ibm.com'
-
-    licenseName = 'The Apache Software License, Version 2.0'
-    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-    allLicenses = ["Apache-2.0"]
-}
-
-android {
-    compileSdkVersion 23
-    buildToolsVersion "21.1.2"
-
-    defaultConfig {
-        // BLE introduced at SDK Version 18
-        minSdkVersion 18
-        targetSdkVersion 23
-        versionCode 5
-        versionName version
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
-        }
-    }
-}
-
-dependencies {
-    compile 'com.android.support:support-v4:23.1.1'
-    compile 'org.altbeacon:android-beacon-library:2.7'
-    compile files('libs/JSON4J.jar')
-}
-
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
