--- pi-sdk/src/main/AndroidManifest.xml
@@ -1,31 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ibm.pisdk" >
-    
-    <uses-permission android:name="android.permission.BLUETOOTH"/>
-	<uses-permission android:name="android.permission.INTERNET" />
-	<uses-permission android:name="android.permission.BLUETOOTH_ADMIN"/>
-	<uses-feature android:name="android.hardware.bluetooth_le" android:required="true"/>
-	<permission android:name="com.ibm.pz.permission.C2D_MESSAGE" android:protectionLevel="signature" />
-	<uses-permission android:name="com.ibm.pz.permission.C2D_MESSAGE" />
-	<uses-permission android:name="com.google.android.c2dm.permission.RECEIVE" />
-	<uses-permission android:name="android.permission.WAKE_LOCK" />
-	<uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
-    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
-
-    <application
-        android:allowBackup="true"
-        android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
-
-        <service android:name=".PIBeaconSensorService"/>
-
-        <receiver android:name="com.ibm.pisdk.onBootComplete">
-            <intent-filter>
-                <action android:name="android.intent.action.BOOT_COMPLETED" />
-                <action android:name="android.intent.action.QUICKBOOT_POWERON" />
-            </intent-filter>
-        </receiver>
-
-    </application>
-
-</manifest>
