--- pi-sdk/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">PISDK</string>
-    <string name="pi_shared_pref">com.ibm.pisdk.shared_prefs</string>
-</resources>
