+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':pi-sdk', ':pi-beacon', ':pi-geofence', ':pi-core'
+include ':pi-beacon', ':pi-geofence', ':pi-core'
